requirements.txt
click==8.1.3
cycler==0.11.0
Flask==2.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.5.2
mysql-connector-python==8.0.30
numpy==1.23.1
pandas==1.4.3
Pillow==9.2.0
protobuf==3.20.1
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
scikit-learn==1.1.1
scipy==1.9.0
seaborn==0.11.2
six==1.16.0
sklearn==0.0
threadpoolctl==3.1.0
Werkzeug==2.2.1
